fixup! fixup! Update cmd/root.go
