+++ test/admission_test.go
@@ -8,7 +8,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/canonical/k8s-dqlite/pkg/kine/endpoint"
 	. "github.com/onsi/gomega"
 	clientv3 "go.etcd.io/etcd/client/v3"
 )
@@ -17,16 +16,7 @@ import (
 // by the admission control.
 func TestAdmissionControl(t *testing.T) {
 	ctx := context.Background()
-	dqlite, err := newDqliteApp(t)
-	if err != nil {
-		t.Fatalf("Failed to create dqlite app: %v", err)
-	}
-	dir := t.TempDir()
-	endpointConfig := endpoint.Config{
-		Listener: fmt.Sprintf("unix://%s/listen.sock", dir),
-		Endpoint: fmt.Sprintf("dqlite://%s/data.db?admission-control-policy=limit&admission-control-policy-limit-max-concurrent-txn=600&driver-name=%s", dir, dqlite.Driver()),
-	}
-	client, _ := newKineWithEndpoint(ctx, t, endpointConfig)
+	client, _ := newKine(ctx, t, "admission-control-policy=limit", "admission-control-policy-limit-max-concurrent-txn=600")
 	g := NewWithT(t)
 
 	// create a key space of 1000 items
